refactor(stream_processor): Extract AgentConfigurator class
This commit extracts all agent and retriever configuration logic from the StreamProcessor class into a new AgentConfigurator class. This significantly improves modularity, reduces the StreamProcessor's responsibilities by breaking down a 'god object', and enhances testability by adhering to the Single Responsibility Principle. The StreamProcessor now orchestrates the configuration process by utilizing the new AgentConfigurator.
